use admin-specific query/mutation
